Working on transactions

+++ includes/model/db/tables/class-usctdp-mgmt-activity-link-table.php
@@ -18,9 +18,9 @@ class Usctdp_Mgmt_Activity_Link_Table extends Table
     {
         $this->schema = "
             id bigint(20) unsigned NOT NULL AUTO_INCREMENT,
-            activity_id int(11) unsigned NOT NULL,
-            session_id int(11) unsigned NOT NULL,
-            clinic_id int(11) unsigned NOT NULL,
+            activity_id bigint(20) unsigned NOT NULL,
+            session_id bigint(20) unsigned NOT NULL,
+            clinic_id bigint(20) unsigned NOT NULL,
             PRIMARY KEY (id),
             KEY activity_id (activity_id),
             KEY session_id (session_id),
